requirements.txt
APScheduler==3.6.3
black==22.3.0
cachetools==4.2.2
certifi==2021.10.8
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
